[XDMAC] Regenerated demo for xdmac_xdmac_memory_transfer_sam_a7g5_ek
MH3-78718

+++ apps/xdmac/xdmac_memory_transfer/firmware/src/config/sam_a7g5_ek/peripheral/pit64b/plib_pit64b0.c
@@ -41,6 +41,7 @@
 
 #include <stddef.h>
 #include "plib_pit64b0.h"
+#include "interrupts.h"
 
 #define PIT64B0_COUNTER_FREQUENCY (200000000U / 1U)
 
@@ -52,12 +53,13 @@ typedef struct
 } PIT64B0_OBJECT;
 
 
-static PIT64B0_OBJECT pit64b0 = 
+static PIT64B0_OBJECT pit64b0 =
 {
     false,
     0U,
     10U,
 };
+
 
 
 static inline void PIT64B0_PERIOD_SET(uint32_t periodLSB, uint32_t periodMSB)
@@ -103,8 +105,8 @@ void PIT64B0_TimerPeriodSet(uint64_t period)
 {
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wlong-long"
-    pit64b0.periodMSB = (period & 0xFFFFFFFF00000000U) >> 32U;
-    pit64b0.periodLSB = (period & 0xFFFFFFFFU);
+    pit64b0.periodMSB = (uint32_t)((period & 0xFFFFFFFF00000000U) >> 32U);
+    pit64b0.periodLSB = (uint32_t)(period & 0xFFFFFFFFU);
 #pragma GCC diagnostic pop
     PIT64B0_PERIOD_SET(pit64b0.periodLSB, pit64b0.periodMSB);
 }
@@ -138,7 +140,7 @@ void PIT64B0_DelayMs(uint32_t delay_ms)
 {
     uint64_t newCount = 0U, deltaCount = 0U, elapsedCount = 0U;
     uint64_t period = PIT64B0_TimerPeriodGet() + 1UL;
-    uint64_t delayCount = (PIT64B0_COUNTER_FREQUENCY / 1000U) * delay_ms;
+    uint64_t delayCount = (PIT64B0_COUNTER_FREQUENCY / 1000U) * (uint64_t)delay_ms;
     uint64_t oldCount = PIT64B0_TimerCounterGet();
     if(pit64b0.running)
     {
@@ -157,7 +159,7 @@ void PIT64B0_DelayUs(uint32_t delay_us)
 {
     uint64_t newCount = 0U, deltaCount = 0U, elapsedCount = 0U;
     uint64_t period = PIT64B0_TimerPeriodGet() + 1UL;
-    uint64_t delayCount = (PIT64B0_COUNTER_FREQUENCY / 1000000U) * delay_us;
+    uint64_t delayCount = (PIT64B0_COUNTER_FREQUENCY / 1000000U) * (uint64_t)delay_us;
     uint64_t oldCount = PIT64B0_TimerCounterGet();
     if(pit64b0.running)
     {
